+++ src/net.c
@@ -75,96 +75,166 @@ const char* proto_name(uint8_t proto)
     return "unresolved";
 }
 
+
 /**
- * @brief   validate IPv4 header checksum
+ * FROM Suricata project - https://github.com/OISF/suricata
+ * EXACT LOCATION - https://github.com/OISF/suricata/blob/suricata-7.0.2/src/decode-ipv4.h#L196-L244C2
  *
- * @param   packet - pointer to beginning of IPv4 packet
+ * @brief Calculate or validate the checksum for the IP packet
  *
- * @return  0 if checksum is valid
+ * @param pkt  Pointer to the start of the IP packet
+ * @param hlen Length of the IP header
+ * @param init The current checksum if validating, 0 if generating.
+ *
+ * @return csum For validation 0 will be returned for success, for calculation
+ *    this will be the checksum.
  */
-uint16_t ipv4_header_checksum(const uint16_t *packet)
+static inline uint16_t ipv4_checksum(const uint16_t *pkt, uint16_t hlen, uint16_t init)
 {
-    //IPv4 header length in bytes
-    uint8_t hdr_length_b = (*packet & 0x0F) * 4; //IHL = number of 32-bit words. IHL * 32 / 8 = IHL * 4
+    uint32_t csum = init;
 
-    uint32_t checksum = 0;
+    csum += pkt[0] + pkt[1] + pkt[2] + pkt[3] + pkt[4] + pkt[6] + pkt[7] +
+        pkt[8] + pkt[9];
 
-    for (uint8_t i = 0; i < hdr_length_b / 2; ++i) {
-        checksum += packet[i];
+    hlen -= 20;
+    pkt += 10;
+
+    if (hlen == 0) {
+        ;
+    } else if (hlen == 4) {
+        csum += pkt[0] + pkt[1];
+    } else if (hlen == 8) {
+        csum += pkt[0] + pkt[1] + pkt[2] + pkt[3];
+    } else if (hlen == 12) {
+        csum += pkt[0] + pkt[1] + pkt[2] + pkt[3] + pkt[4] + pkt[5];
+    } else if (hlen == 16) {
+        csum += pkt[0] + pkt[1] + pkt[2] + pkt[3] + pkt[4] + pkt[5] + pkt[6] +
+            pkt[7];
+    } else if (hlen == 20) {
+        csum += pkt[0] + pkt[1] + pkt[2] + pkt[3] + pkt[4] + pkt[5] + pkt[6] +
+            pkt[7] + pkt[8] + pkt[9];
+    } else if (hlen == 24) {
+        csum += pkt[0] + pkt[1] + pkt[2] + pkt[3] + pkt[4] + pkt[5] + pkt[6] +
+            pkt[7] + pkt[8] + pkt[9] + pkt[10] + pkt[11];
+    } else if (hlen == 28) {
+        csum += pkt[0] + pkt[1] + pkt[2] + pkt[3] + pkt[4] + pkt[5] + pkt[6] +
+            pkt[7] + pkt[8] + pkt[9] + pkt[10] + pkt[11] + pkt[12] + pkt[13];
+    } else if (hlen == 32) {
+        csum += pkt[0] + pkt[1] + pkt[2] + pkt[3] + pkt[4] + pkt[5] + pkt[6] +
+            pkt[7] + pkt[8] + pkt[9] + pkt[10] + pkt[11] + pkt[12] + pkt[13] +
+            pkt[14] + pkt[15];
+    } else if (hlen == 36) {
+        csum += pkt[0] + pkt[1] + pkt[2] + pkt[3] + pkt[4] + pkt[5] + pkt[6] +
+            pkt[7] + pkt[8] + pkt[9] + pkt[10] + pkt[11] + pkt[12] + pkt[13] +
+            pkt[14] + pkt[15] + pkt[16] + pkt[17];
+    } else if (hlen == 40) {
+        csum += pkt[0] + pkt[1] + pkt[2] + pkt[3] + pkt[4] + pkt[5] + pkt[6] +
+            pkt[7] + pkt[8] + pkt[9] + pkt[10] + pkt[11] + pkt[12] + pkt[13] +
+            pkt[14] + pkt[15] + pkt[16] + pkt[17] + pkt[18] + pkt[19];
     }
 
-    checksum = (checksum >> 16) + (checksum & 0x0000FFFF);
+    csum = (csum >> 16) + (csum & 0x0000FFFF);
+    csum += (csum >> 16);
 
-    return (uint16_t) ~checksum;
+    return (uint16_t) ~csum;
+}
+
+static inline uint8_t handle_ipv4(ipv4_h *ipv4, packet *packet)
+{
+    packet->network_l.h_len = (ipv4->verhl & 0x0F) * 4;
+
+    if (ipv4_checksum((uint16_t*)ipv4, packet->network_l.h_len, ipv4->checksum) == SUCCESS) {
+
+        packet->stack |= IPV4;
+        packet->network_l.header = (void*)ipv4;
+
+        if (ipv4->next_proto == TCP) {
+            packet->stack |= TCP;
+            packet->transport_l.header = ((void*)ipv4) + packet->network_l.h_len;
+            packet->transport_l.h_len = (((tcp_h*)packet->transport_l.header)->data_offset >> 4) * 4;
+        }
+
+        return SUCCESS;
+    }
+    return FAIL;
+}
+
+static inline uint8_t handle_ipv6(ipv6_h *ipv6, packet *packet)
+{
+    if (ipv6->ver >> 4 == IPv6_VER) {
+
+        packet->stack |= IPV6;
+        packet->network_l.h_len = IPv6_HDR_LEN;
+        packet->network_l.header = (void*)ipv6;
+
+        if (ipv6->next_proto == TCP) {
+            packet->stack |= TCP;
+            packet->transport_l.header = (void*)(ipv6 + 1);
+            packet->transport_l.h_len = (((tcp_h*)packet->transport_l.header)->data_offset >> 4) * 4;
+        }
+
+        return SUCCESS;
+    }
+    return FAIL;
 }
 
 /**
- * @brief   parse packet to respective headers
+ * @brief   thread-safe parse packet to respective layers
  *
- * @param   hdrs - struct of founded headers
- * @param   pkt  - pointer to beginning of some packet
- *
- * @return  IP version, 0 if not discovered
+ * @param   packet - struct of founded headers by layers
  */
-uint8_t protocol_recognition(headers *hdrs, const uint8_t *pkt)
+void stack_recognition(packet *packet)
 {
-    uint8_t ver = *pkt >> 4;
+    /*
+     *  Assume that most common scenario will be capturing from tun* device or NFLOG source
+     *  and most common protocol is IPv4. In this case packet starts with IPv4 header
+     */
+    uint8_t ver = *(uint8_t*)packet->raw_pkt_ptr >> 4;
 
-    if (ver == IPv4_VER) {
-        if (ipv4_header_checksum((uint16_t*)pkt) == 0) {
-            hdrs->ipv4 = (ipv4_h*)pkt;
-            return IPv4_VER;
-        }
+    if (ver == IPv4_VER && handle_ipv4((ipv4_h*)packet->raw_pkt_ptr, packet) == SUCCESS) {
+        return;
+    } //First header not an IPv4 for sure
+
+    packet->link_l.header = packet->raw_pkt_ptr; //put link layer to stack, assuming it either Ethernet or 802.1Q
+    packet->link_l.h_len = ETH_HDR_LEN;
+    packet->stack |= ETH;
+
+    ether_h *eth = (ether_h*)packet->raw_pkt_ptr;
+
+    if (eth->etype == 0x0008 && handle_ipv4((ipv4_h*)(eth + 1), packet) == SUCCESS) { //0x0800 (IPv4) - network order
+        return;
     }
 
-    hdrs->ethernet = (ether_h*)pkt;
-
-    if (hdrs->ethernet->etype == 0x0008) { //0x0800 (IPv4) - network order
-        if (ipv4_header_checksum((uint16_t*)(pkt + ETH_OFFSET)) == 0) {
-            hdrs->ipv4 = (ipv4_h*)(pkt + ETH_OFFSET);
-            return IPv4_VER;
-        }
+    if (eth->etype == 0xDD86 && handle_ipv6((ipv6_h*)(eth + 1), packet) == SUCCESS) { //0x86DD (IPv6) - network order
+        return;
     }
 
-    if (hdrs->ethernet->etype == 0xDD86) { //0x86DD (IPv6) - network order
-        if (pkt[ETH_OFFSET] >> 4 == IPv6_VER) {
-            hdrs->ipv6 = (ipv6_h*)(pkt + ETH_OFFSET);
-            return IPv6_VER;
-        }
-    }
+    if (eth->etype == 0x0081) { //0x8100 (802.1Q) - network order
+        packet->stack &= ~LINK_MASQ;
+        packet->link_l.h_len = DOT1Q_HDR_LEN;
+        packet->stack |= DOT1Q;
+        dot1q_h *dot1q = (dot1q_h*)packet->raw_pkt_ptr;
 
-    if (hdrs->ethernet->etype == 0x0081) { //0x8100 (802.1Q) - network order
-
-        if (hdrs->ethernet->vlan_etype == 0x0008) {
-            if (ipv4_header_checksum((uint16_t*)(pkt + VLAN_OFFSET)) == 0) {
-                hdrs->ipv4 = (ipv4_h*)(pkt + VLAN_OFFSET);
-                return IPv4_VER;
-            }
+        if (dot1q->etype == 0x0008 && handle_ipv4((ipv4_h*)(dot1q + 1), packet) == SUCCESS) {
+            return;
         }
 
-        if (hdrs->ethernet->vlan_etype == 0xDD86) {
-            if (pkt[VLAN_OFFSET] >> 4 == IPv6_VER) {
-                hdrs->ipv6 = (ipv6_h*)(pkt + VLAN_OFFSET);
-                return IPv6_VER;
-            }
+        if (dot1q->etype == 0xDD86 && handle_ipv6((ipv6_h*)(dot1q + 1), packet) == SUCCESS) {
+            return;
         }
+
     }
 
 
     if (ver == IPv6_VER) {
+        ipv6_h *ipv6 = (ipv6_h*)packet->raw_pkt_ptr;
 
-        hdrs->ipv6 = (ipv6_h*)pkt;
-
-        /*
-         * Since IPv6 header hasn't a checksum, the best what we can do
-         * is guessing based on common encapsulated protocol ID.
-         *
-         * Or implement full stack recognition with checksums verification on each layer,
-         * which is great amount of efforts (not for this project anyway)
-         *
-         * If you have any suggestions, please let me know
-         */
-        switch (hdrs->ipv6->next_proto) {
+        switch (ipv6->next_proto) {
+            case 6:     //TCP
+                packet->stack |= TCP;
+                packet->transport_l.header = (void*)(ipv6 + 1);
+                packet->transport_l.h_len = (((tcp_h*)packet->transport_l.header)->data_offset >> 4) * 4;
+                /* no break */
             case 41:    //IPv6
             case 43:    //IPv6-Route
             case 59:    //IPv6-NoNxt
@@ -172,13 +242,15 @@ uint8_t protocol_recognition(headers *hdrs, const uint8_t *pkt)
             case 58:    //IPv6-ICMP
             case 44:    //IPv6-Frag
             case 17:    //UDP
-            case 6:     //TCP
-                hdrs->ethernet = NULL;
-                return IPv6_VER;
-            default:
-                hdrs->ipv6 = NULL;
+                packet->stack |= IPV6;
+                packet->network_l.header = packet->raw_pkt_ptr;
+                packet->network_l.h_len = IPv6_HDR_LEN;
+                packet->stack &= ~LINK_MASQ;
+                packet->link_l.header = NULL;
+                packet->link_l.h_len = 0;
+
         }
+
     }
 
-    return 0;
 }