+++ src/tuninetd.c
@@ -50,11 +50,13 @@ int main(int argc, char *argv[])
 void build_config(int argc, char **argv)
 {
     int opt = 0;
-    static const char *optString = "i:t:c:f:m:n:dhv";
+    static const char *optString = "i:t:c:f:m:n:w:adhv";
 
+    globcfg.ack_only = 0;
     globcfg.isdaemon = 0;
     globcfg.pid = 0;
     globcfg.cmd_path = NULL;
+    globcfg.pcap_file_path = NULL;
     globcfg.ttl = 600;
     globcfg.nf_group = -1;
 
@@ -62,6 +64,9 @@ void build_config(int argc, char **argv)
 
     while( opt != -1 ) {
         switch( opt ) {
+            case 'a':
+                globcfg.ack_only = 1;
+                break;
             case 'v':
                 version();
                 exit(0);
@@ -74,15 +79,17 @@ void build_config(int argc, char **argv)
             case 'c':
                 globcfg.cmd_path = optarg;
 
-                globcfg.cmd_path_start = malloc(strlen(optarg) + 23);
+                globcfg.cmd_path_start = malloc(strlen(optarg) + 24);
                 strcpy(globcfg.cmd_path_start, optarg);
                 strcat(globcfg.cmd_path_start, " start > /dev/null 2>&1");
 
-                globcfg.cmd_path_stop = malloc(strlen(optarg) + 22);
+                globcfg.cmd_path_stop = malloc(strlen(optarg) + 23);
                 strcpy(globcfg.cmd_path_stop, optarg);
                 strcat(globcfg.cmd_path_stop, " stop > /dev/null 2>&1");
                 break;
-
+            case 'w':
+                globcfg.pcap_file_path = optarg;
+                break;
             case 'f':
                 globcfg.pcap_filter = optarg;
                 break;
@@ -108,13 +115,13 @@ void build_config(int argc, char **argv)
 void check_config_and_daemonize()
 {
     if (globcfg.dev_name == NULL && globcfg.nf_group < 0) {
-        message(ERROR, "Network device or nflog-group must be specified. Abort.");
+        message(ERROR, "Network device or nflog-group must be specified. Abort");
         usage();
         exit(1);
     }
 
     if (globcfg.cmd_path == NULL) {
-        message(ERROR, "Executable path must be specified. Abort.");
+        message(ERROR, "Executable path must be specified. Abort");
         usage();
         exit(1);
     }
@@ -123,7 +130,7 @@ void check_config_and_daemonize()
         globcfg.pid = fork();
 
         if (globcfg.pid < 0) {
-            message(ERROR, "Can't fork process. Abort.");
+            message(ERROR, "Can't fork process. Abort");
             exit(1);
         }
 
@@ -133,7 +140,9 @@ void check_config_and_daemonize()
             exit(0);
         }
 
-        chdir("/");
+        if (chdir("/") < 0) {
+            message(WARNING, "Can't change directory");
+        }
 
         setsid();
 
@@ -153,11 +162,11 @@ void version() {
 void sighup_handler(int signo)
 {
     if (status == OFF) {
-       message(WARNING, "Warning! Tuninetd is already in standby mode.");
+       message(WARNING, "Warning! Tuninetd is already in standby mode");
        return;
     }
 
-    message(INFO, "SIGHUP caught, switch to standby mode.");
+    message(INFO, "SIGHUP caught, switch to standby mode");
 
     switch_guard(OFF);
 }
@@ -169,24 +178,32 @@ void sigusr_handler(int signo)
     message(INFO, "SIGUSR1 caught:");
 
     if (globcfg.dev_name) {
-        message(INFO, "- Capture engine: pcap, %s", globcfg.dev_name);
+        message(INFO, "- capture engine: pcap, %s", globcfg.dev_name);
         if (globcfg.pcap_filter != NULL) {
-            message(INFO, "-- Pcap filter: \"%s\"", globcfg.pcap_filter);
+            message(INFO, "-- pcap filter: \"%s\"", globcfg.pcap_filter);
         }
     }
 
     if (globcfg.nf_group >= 0) {
-        message(INFO, "- Capture engine: nflog group %ld", globcfg.nf_group);
+        message(INFO, "- capture engine: nflog group %ld", globcfg.nf_group);
     }
 
-    message(INFO, "- cmd_path = %s", globcfg.cmd_path);
-    message(INFO, "- TTL = %ld sec.", globcfg.ttl);
+    if (globcfg.ack_only == ON) {
+        message(INFO, "- using packets with tcp[ack] flag to reset TTL timer");
+    }
+
+    if (globcfg.pcap_file_path != NULL) {
+        message(INFO, "- file to dump start event packets (pcap format): %s", globcfg.pcap_file_path);
+    }
+
+    message(INFO, "- event path: %s", globcfg.cmd_path);
+    message(INFO, "- TTL: %ld sec.", globcfg.ttl);
 
     if (status == OFF) {
-        message(INFO, "- Current status: standby (OFF)");
+        message(INFO, "- current status: standby (OFF)");
     } else {
         delta = curts - ts;
-        message(INFO, "- Current status: up (ON), time since last captured packet: %ld sec.", delta < 0 ? 0 : delta);
+        message(INFO, "- current status: up (ON), time since last captured packet: %ld sec.", delta < 0 ? 0 : delta);
     }
 }
 
@@ -202,20 +219,22 @@ void sigterm_handler(int signo)
 void usage(void) {
     fprintf(stderr, VERSION);
     fprintf(stderr, "\nEvent emitter with pcap and nflog sensors which could be used at the same time\n");
-    fprintf(stderr, "\nUsage: \t%s -i <ifname> -c <path> [-f <filter>] [-n <nflog-group>] [-t <ttl>] [-d]\n", progname);
-    fprintf(stderr, "\t%s -n <nflog-group> -c <path> [-i <ifname> [-f <filter>]] [-t <ttl>] [-d]\n", progname);
+    fprintf(stderr, "\nUsage: \t%s -i <ifname> -c <path> [-a] [-d] [-f <filter>] [-n <nflog-group>] [-t <ttl>] [-w <path>]\n", progname);
+    fprintf(stderr, "\t%s -n <nflog-group> -c <path> [-a] [-d] [-i <ifname> [-f <filter>]] [-t <ttl>] [-w <path>]\n", progname);
     fprintf(stderr, "\n\n");
-    fprintf(stderr, "-i <ifname>: network interface to use with pcap. Must be up and configured.\n");
-    fprintf(stderr, "-c <path>: to executable, will be run with 'start' and 'stop' parameter accordingly.\n");
-    fprintf(stderr, "-n <nflog-group>: netfilter nflog group number (0 - 65535)\n");
-    fprintf(stderr, "-f <filter>: specify pcap filter if needed, similar to tcpdump. Default none (all packets)\n");
-    fprintf(stderr, "-t <ttl>: seconds of interface idle before 'stop' command will be run. Default 600.\n");
-    fprintf(stderr, "-d: daemonize process. Check for errors before use.\n\n");
+    fprintf(stderr, "-a: use only tcp[ack] packets to zero TTL timer (see -t)\n");
+    fprintf(stderr, "-c <path>: to executable, will be run with 'start' and 'stop' parameter accordingly\n");
+    fprintf(stderr, "-d: daemonize process\n");
+    fprintf(stderr, "-f <filter>: apply packet filter, similar to tcpdump\n");
+    fprintf(stderr, "-i <ifname>: network interface to use with pcap\n");
+    fprintf(stderr, "-n <nflog-group>: netfilter nflog group ID\n");
+    fprintf(stderr, "-t <ttl>: seconds of interface idle before 'stop' command will be run. Default 600\n");
+    fprintf(stderr, "-w <path>: dump \"start event\" packets to pcap-savefile as well\n\n");
     fprintf(stderr, "-h: print this help\n\n");
     fprintf(stderr, "-v: print version\n\n");
     fprintf(stderr, "\nExamples:\n\n");
-    fprintf(stderr, "# tuninetd -n 1 -c /path/to/executable/toggletunnel.sh\n");
-    fprintf(stderr, "# tuninetd -i tap0 -c /path/to/executable/coolscript.sh\n");
+    fprintf(stderr, "# tuninetd -n 1 -c /path/to/executable/toggletunnel.sh -w /path/to/pcap-savefile\n");
+    fprintf(stderr, "# tuninetd -i tap0 -c /path/to/executable/coolscript.sh -a\n");
     fprintf(stderr, "# tuninetd -i tun0 -f \"! host 1.2.3.4\" -c /path/to/executable/binary -t 3600 -d\n");
     fprintf(stderr, "# tuninetd -i enp3s0 -f \"arp and host 4.3.2.1\" -n 1 -c /path/to/executable/launcher.py\n\n");
     fprintf(stderr, "More information: https://github.com/root4root/tuninetd \n\n");